+++ packages/react/src/box/index.tsx
@@ -1,6 +1,6 @@
 import "@postenbring/hedwig-css/dist/fonts.css";
 import "@postenbring/hedwig-css/dist/tokens.css";
-import "@postenbring/hedwig-css/dist/base.css";
+import "@postenbring/hedwig-css/dist/body.css";
 
 import "@postenbring/hedwig-css/dist/box.css";