+++ packages/react/src/dropdown/index.tsx
@@ -2,6 +2,6 @@ import "@postenbring/hedwig-css/dist/fonts.css";
 import "@postenbring/hedwig-css/dist/tokens.css";
 import "@postenbring/hedwig-css/dist/body.css";
 
-import "@postenbring/hedwig-css/dist/link.css";
+import "@postenbring/hedwig-css/dist/dropdown.css";
 
-export * from "./link";
+export * from "./dropdown";