+++ packages/tokens/tailwind.config.cjs
@@ -33,51 +33,91 @@ function expandAndPrefix(obj, prefix) {
 
 /**
  * @param {keyof typeof tokens["font-size"]} name
+ * @param {`var(--hds-${string})` | keyof typeof tokens["font-weight"]} fontWeightName
  * @returns
  */
-function fontSize(name) {
-  return [tokens["font-size"][name], tokens["line-height"][name]];
+function fontSize(name, fontWeightName) {
+  let fontWeight;
+  if (fontWeightName.startsWith("var(--hds-")) {
+    fontWeight = fontWeightName;
+  } else {
+    // @ts-ignore -- It's ok, it's matched in the function signature
+    fontWeight = tokens["font-weight"][fontWeightName];
+  }
+  return [
+    tokens["font-size"][name],
+    {
+      lineHeight: tokens["line-height"][name],
+      fontWeight,
+    },
+  ];
 }
 
 /** @type {import('tailwindcss').Config} */
 module.exports = {
   theme: {
     fontFamily: {
-      bold: tokens.fonts["posten-sans-bold"],
-      medium: tokens.fonts["posten-sans-medium"],
-      regular: tokens.fonts["posten-sans-regular"],
-      light: tokens.fonts["posten-sans-light"],
+      DEFAULT: tokens.fonts["posten-sans"],
+    },
+    fontWeight: {
+      regular: 400,
+      medium: 500,
+      bold: 700,
+      light: 300,
     },
     fontSize: {
-      "h1-display": fontSize("posten-h1-display"),
-      "h1-display-min": fontSize("posten-h1-display-min"),
-      "h1-display-max": fontSize("posten-h1-display-max"),
+      "h1-display": fontSize("posten-h1-display", "var(--hds-brand-font-weight-h1-display)"),
+      "h1-display-min": fontSize(
+        "posten-h1-display-min",
+        "var(--hds-brand-font-weight-h1-display)",
+      ),
+      "h1-display-max": fontSize(
+        "posten-h1-display-max",
+        "var(--hds-brand-font-weight-h1-display)",
+      ),
 
-      h1: fontSize("posten-h1"),
-      "h1-min": fontSize("posten-h1-min"),
-      "h1-max": fontSize("posten-h1-max"),
+      h1: fontSize("posten-h1", "var(--hds-brand-font-weight-h1)"),
+      "h1-min": fontSize("posten-h1-min", "var(--hds-brand-font-weight-h1)"),
+      "h1-max": fontSize("posten-h1-max", "var(--hds-brand-font-weight-h1)"),
 
-      h2: fontSize("posten-h2"),
-      "h2-min": fontSize("posten-h2-min"),
-      "h2-max": fontSize("posten-h2-max"),
+      h2: fontSize("posten-h2", "var(--hds-brand-font-weight-h2)"),
+      "h2-min": fontSize("posten-h2-min", "var(--hds-brand-font-weight-h2)"),
+      "h2-max": fontSize("posten-h2-max", "var(--hds-brand-font-weight-h2)"),
 
-      h3: fontSize("header-h3"),
-      "h3-min": fontSize("header-h3-min"),
-      "h3-max": fontSize("header-h3-max"),
+      h3: fontSize("header-h3", "header-h3"),
+      "h3-min": fontSize("header-h3-min", "header-h3"),
+      "h3-max": fontSize("header-h3-max", "header-h3"),
 
-      body: fontSize("body"),
-      "body-min": fontSize("body-min"),
-      "body-max": fontSize("body-max"),
+      "h3-title": fontSize("header-h3-title", "header-h3-title"),
+      "h3-title-min": fontSize("header-h3-title-min", "header-h3-title"),
+      "h3-title-max": fontSize("header-h3-title-max", "header-h3-title"),
 
-      "body-small": fontSize("body small"),
-      "body-small-min": fontSize("body small-min"),
-      "body-small-max": fontSize("body small-max"),
+      body: fontSize("body", "body"),
+      "body-min": fontSize("body-min", "body"),
+      "body-max": fontSize("body-max", "body"),
 
-      technical: fontSize("technical"),
-      "technical-min": fontSize("technical-min"),
-      "technical-max": fontSize("technical-max"),
+      "body-title": fontSize("body title", "body title"),
+      "body-title-min": fontSize("body title-min", "body title"),
+      "body-title-max": fontSize("body title-max", "body title"),
 
-      caption: fontSize("caption"),
+      "body-small": fontSize("body small", "body small"),
+      "body-small-min": fontSize("body small-min", "body small"),
+      "body-small-max": fontSize("body small-max", "body small"),
+
+      "body-small-title": fontSize("body small title", "body small title"),
+      "body-small-title-min": fontSize("body small title-min", "body small title"),
+      "body-small-title-max": fontSize("body small title-max", "body small title"),
+
+      technical: fontSize("technical", "technical"),
+      "technical-min": fontSize("technical-min", "technical"),
+      "technical-max": fontSize("technical-max", "technical"),
+
+      "technical-title": fontSize("technical title", "technical title"),
+      "technical-title-min": fontSize("technical title-min", "technical title"),
+      "technical-title-max": fontSize("technical title-max", "technical title"),
+
+      caption: fontSize("caption", "caption"),
+      "caption-title": fontSize("caption title", "caption title"),
     },
     colors: {
       // The brand specifc tokens are themeable,
@@ -127,7 +167,6 @@ module.exports = {
     extend: {},
   },
   corePlugins: {
-    fontWeight: false,
     ringColor: false,
     ringOffsetColor: false,
     ringOffsetWidth: false,